Delete GTU_REPORT.pdf
This REPORT gives an insight into the research and agenda behind the project and idea. It also describes the overall process in detail including the knowledge of PLC, Code for the project(Ladder Logic), Electrical Wiring Diagram, Mechanical Design, Mathematical Calculations of all the components used to design the system.